refactor: Better physical file organization
