+++ src/device-interfaces/i2c/hw-i2c.ts
