+++ src/device-interfaces/i2c/i2c-connection.ts
