+++ src/device-interfaces/i2c/i2c-error-detector.ts
