+++ src/device-interfaces/i2c/mock-i2c-device.ts
